Install binary to sbin and don't abort the program if a wrong input is encountered.
As previously with autotools, we want to install the pxe-kexec binary to sbin.

Also don't abort the program when a wrong input is encountered in the question
that asks the user if he really wants to load the kernel or edit the
commandline. Instead, just re-ask.

committer: Bernhard Walle <bwalle@suse.de>