Improve error handling when connection fails.
When the connection fails to the server when fetching the PXE configuration
files, pxe-kexec now doesn't try to fetch other files because that also
will fail.

When the TFTP protocol is used to fetch the PXE configuration files (which is
the default), then it also prints a nice error message which gives the user the
hint to check the firewall configuration. The message also mentions the
'-F' option to use FTP instead of TFTP which was implemented because of exactly
that reason.

Because all of this, a timeout of 7 seconds was added.

committer: Bernhard Walle <bwalle@suse.de>